requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
openai==1.35.4
google-generativeai==0.7.1
python-multipart==0.0.9
pandas==2.2.2
